pyproject.toml
[project]
name = "diffusion2"
version = "0.1.0"
description = "Minimal Diffusion Model full dev"
readme = "README.md"
authors = [{name="Shubham"},]
requires-python = ">=3.11"
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "pydantic",
    "pyyaml",
    "tqdm",
    "uvicorn",
    "fastapi",
    "pillow"
]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"